+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+v 1.5.0
+
++ Balance Top command (/balancetop, /baltop)
++ Added alias for the set balance command (/setbal)
++ Added a money cap that can be enabled/disabled as well as set from the main config
++ Ability to set default job notifications state. Thanks to @MarkL4YG.
++ Jobs config can now be reloaded without having to restart the server. (/sponge plugins reload)
++ The pay command can be used to add/remove money from the console. To remove money just do "pay [NAME] -[AMOUNT]".
++ Job Info command (/job info) (Displays a paginated list of everything that earns exp/money for the players current job in a paginated list)
++ Updated to latest versions of Forge and Sponge
++ CHANGE: EXP/Money is no longer rewarded for breaking blocks that a player placed. This replaces the preventJobFarming node.
++ FIX: Job permissions were changed back to their normal values
++ FIX: Issue with Cause.of has been fixed
+
+v 1.4.0
+
++ Updated API version
+
 v 1.3.0
 
 + Updated to use the Economy API and to work with the latest version of Sponge